TASK: Split glorified single interface haystack into `@neos-project/neos-ui-contentrepository-model`
