+++ packages/neos-ui-backend-connector/package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@neos-project/neos-ts-interfaces": "workspace:*",
+    "@neos-project/neos-ui-contentrepository-model": "workspace:*",
     "@neos-project/utils-helpers": "workspace:*"
   },
   "license": "GNU GPLv3"
