+++ packages/neos-ui-backend-connector/src/FlowQuery/Operations/Context/index.ts
@@ -1,4 +1,4 @@
-import {ContextProperties} from '@neos-project/neos-ts-interfaces';
+import {ContextProperties} from '@neos-project/neos-ui-contentrepository-model';
 
 export default () => (contextProperties: ContextProperties) => ({
     type: 'context',
