+++ packages/neos-ui-backend-connector/src/FlowQuery/Operations/Search/index.ts
@@ -1,4 +1,4 @@
-import {NodeTypeName} from '@neos-project/neos-ts-interfaces';
+import {NodeTypeName} from '@neos-project/neos-ui-contentrepository-model';
 
 export default () => (term: string, filterNodeType: NodeTypeName) => ({
     type: 'search',
