+++ packages/neos-ui-contentrepository-model/src/dimension/index.ts
@@ -1,0 +1,2 @@
+
+export * from './dimension';
