+++ packages/neos-ui-contentrepository-model/src/index.ts
@@ -1,1 +1,5 @@
 export * from './node';
+export * from './nodetype';
+export * from './workspace';
+export * from './dimension';
+export * from './subgraph';
