+++ packages/neos-ui-contentrepository-model/src/node/index.ts
@@ -1,1 +1,2 @@
 export {PropertyScope} from './propertyScope';
+export * from './node';
