+++ packages/neos-ui-contentrepository-model/src/nodetype/index.ts
@@ -1,0 +1,5 @@
+
+export {NodeType} from './nodetype';
+export {PropertyConfiguration} from './properties';
+export {ReferencesConfiguration} from './references';
+export {ValidatorConfiguration} from './validators';
