+++ packages/neos-ui-contentrepository-model/src/nodetype/validators.ts
@@ -1,0 +1,4 @@
+
+export interface ValidatorConfiguration {
+    [propName: string]: any;
+}
