+++ packages/neos-ui-contentrepository-model/src/subgraph/index.ts
@@ -1,0 +1,1 @@
+export {ContextProperties} from './subgraph';
