+++ packages/neos-ui-contentrepository-model/src/subgraph/subgraph.ts
@@ -1,0 +1,10 @@
+import {WorkspaceName} from '../workspace';
+import {NodeContextPath} from '../node';
+
+// TODO Rename to SubgraphAddress and align to Neos ESCR model
+export interface ContextProperties {
+    contextPath?: NodeContextPath;
+    workspaceName?: WorkspaceName;
+    invisibleContentShown?: boolean;
+    removedContentShown?: boolean;
+}
