+++ packages/neos-ui-contentrepository-model/src/workspace/index.ts
@@ -1,0 +1,3 @@
+
+export {WorkspaceName, Workspace} from './workspace';
+export {WorkspaceStatus} from './status';
