+++ packages/neos-ui-contentrepository-model/src/workspace/status.ts
@@ -1,0 +1,5 @@
+
+export enum WorkspaceStatus {
+    UP_TO_DATE = 'UP_TO_DATE',
+    OUTDATED = 'OUTDATED'
+}
