+++ packages/neos-ui-contentrepository-model/src/workspace/workspace.ts
@@ -1,0 +1,7 @@
+export type WorkspaceName = string;
+
+export interface Workspace {
+    name: WorkspaceName;
+    title: string;
+    readonly: boolean;
+}
