+++ packages/neos-ui-redux-store/src/CR/ContentDimensions/index.ts
@@ -5,7 +5,7 @@ import {action as createAction, ActionType} from 'typesafe-actions';
 import {actionTypes as system, InitAction} from '../../System';
 import {GlobalState} from '../..';
 import isEqual from 'lodash.isequal';
-import {DimensionInformation, DimensionCombination, DimensionPresetName, DimensionPresetCombination} from '@neos-project/neos-ts-interfaces';
+import {DimensionInformation, DimensionCombination, DimensionPresetName, DimensionPresetCombination} from '@neos-project/neos-ui-contentrepository-model';
 
 //
 // Export the subreducer state shape interface
