+++ packages/neos-ui-sagas/src/CR/NodeOperations/reloadNodes.ts
@@ -11,7 +11,7 @@ import {call, put, select} from 'redux-saga/effects';
 
 import {actions, selectors} from '@neos-project/neos-ui-redux-store';
 import {GlobalState} from '@neos-project/neos-ui-redux-store';
-import {DimensionCombination, Node, NodeContextPath, NodeMap, WorkspaceName} from '@neos-project/neos-ts-interfaces';
+import {DimensionCombination, Node, NodeContextPath, NodeMap, WorkspaceName} from '@neos-project/neos-ui-contentrepository-model';
 import {AnyError} from '@neos-project/neos-ui-error';
 import backend from '@neos-project/neos-ui-backend-connector';
 // @ts-ignore
